// backend/src/enums/sortField.enum.ts
export enum SortField {
    NAME = 'name',
    SURNAME = 'surname',
    EMAIL = 'email',
    AGE = 'age',
    COURSE_NAME = 'courseName',
    COURSE_FORMAT = 'courseFormat',
    COURSE_TARIFF = 'courseTariff',
    GROUP_NAME = 'groupName',
    PHONE = 'phone',
    STATUS = 'status',
    ALREADY_PAID = 'alreadyPaid',
}

export enum OrderDirection {
    ASC = 'asc',
    DESC = 'desc',
}